requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pydantic>=2.0.0
python-multipart>=0.0.9
openai>=1.0.0
pydantic-settings>=2.0.0
python-docx>=1.1.0
matplotlib>=3.8.0
google-cloud-storage>=2.14.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
